Create input.cpp
Just run the code in the same directory as the output file or just change the path directory in the code. I also overloaded the output operator so that we can be able to verify if the data we have in our linked list is correct. You'll update me if I missed anything