Update QueueLinkList.h
So decided to implement the queue as a doubly linked list. I did not change the names of the function as David asked as of yet... Well I'll do so or he could just do it himself after debugging the code. Here's a checklist of what you requested:
1. private member queueSize 
2. int getQueueSize()
3. queueSize initialisez to zero
4. quequeSize is incremented when queue() function is called and decremented when dequeue() is called

PLEASE NOTE THAT I COULD HAVE MISTAKENLY CHANGED VARIABLE NAMES THAT I WASN'T SUPPOSED TO CHANGE.
( I will leave a "test.cpp" code for you guys to run (compile it at the same directory as output.txt). I have overrided the output operator so that you can see that it works, in case it gets broken :[)